feat(pr, auth): Refactor OAuth device flow and optimize PR UI
Consolidates GitHub and Azure device authorization grant logic into a shared
`useDeviceFlowAuth` composable, improving code reuse and maintainability for
authentication workflows.

Enhances PR list performance by parallelizing diff calculations for additions
and deletions.

Optimizes markdown rendering for PR descriptions and comments by pre-rendering
HTML and centralizing external link handling, reducing re-renders and improving
UI responsiveness.

Standardizes checks for merge conflicts across PR components for consistency.
